Rename server_sort_workshop_table.jsx to ServerSideWorkshopTable.jsx
